Move context folder to state folder
